Move remaing crypto components to application source tree. #3
