+++ esp32/main/CMakeLists.txt
@@ -5,6 +5,8 @@ idf_component_register(
         "wifi/manager.cpp"
         "tasks/currency_update.cpp"
         "containers/crypto.cpp"
+        "lib/crypto/coin_gecko.cpp"
+        "lib/hal/requests.cpp"
         "lib/hal/sntp.cpp"
     INCLUDE_DIRS 
         "."
